% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO.next.R
\name{CFO.next}
\alias{CFO.next}
\title{Determination of the dose level for next cohort}
\usage{
CFO.next(phi, cys, cns, cover.doses=c(0,0,0), curDose,
       add.args=list(alp.prior=phi,bet.prior=1-phi), seed=NULL)
}
\arguments{
\item{phi}{the target DLT rate}

\item{cys}{the current number of DLTs observed in patients for the left, current, and right dose levels.}

\item{cns}{the current number of patients for the left, current, and right dose levels.}

\item{cover.doses}{whether the dose level (left, current and right) is over-toxic or not.
The value is set as 1 if the dose level is overly toxicity; otherwise, it is set to 0.}

\item{curDose}{the current dose level.}

\item{add.args}{additional parameters, usually set as list(alp.prior=phi, bet.prior=1-phi) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seed}{an integer to set as the seed of the random number generator for reproducible results.}
}
\value{
The \code{CFO.next()} function returns a list object comprising the following elements: the target DLT
rate ($target), the current counts of DLTs and patients for the left, current, and right dose levels ($cys and $cns),
the decision for whether to move to the left or right dose level for the next cohort ($decision), and the current
dose after movement ($curDose).
}
\description{
Use the function to determine the dose movement based on the toxicity outcomes of the enrolled cohorts.
}
\details{
The CFO design determines the dose level for the next cohort by assessing evidence from the current
dose level and its adjacent levels. This evaluation is based on odds ratios denoted as \eqn{O_k}, where
k = L, C, R represents left, current, and right dose levels. Additionally, we define \eqn{\overline{O}_k = 1/O_k}.
The ratio \eqn{O_C / \overline{O}_{L}} indicates the inclination for de-escalation, while \eqn{\overline{O}_C / O_R}
quantifies the tendency for escalation. Threshold values \eqn{\gamma_L} and \eqn{\gamma_R} are chosen to
minimize the probability of making incorrect decisions.The decision process is summarized in Table 1
of Jin and Yin (2022).
An overdose control rule is implemented to ensure patient safety. If the data suggest excessive
toxicity at the current dose level, we exclude that level and those higher levels. Two scenarios
lead to a decision on one side only: when the current dose is at the boundary (the first or last dose level)
or when higher dose levels have been eliminated.
}
\note{
The \code{cover.doses()} should be predetermined. It can be calculated based on the number of patients enrolled
($cns) and the number of DLTs observed ($cys), following the overdose control rule. If not predetermined, it is
assigned default values of 0 for each dose level.
}
\examples{
## determine the dose level for the next cohort of new patients
cys <- c(0,1,0); cns <- c(3,6,0)
CFO.next(phi=0.2, cys=cys, cns=cns, cover.doses=c(0,0,0), curDose=3,
         add.args=list(alp.prior=0.2, bet.prior=0.8))

}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066.
}
\author{
Jialu Fang and Wenliang Wang
}
