% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO.oc.R
\name{CFO.oc}
\alias{CFO.oc}
\title{Generate operating characteristics for multiple simulations}
\usage{
CFO.oc(nsimu=5000, design, phi, p.true, ncohort, init.level, cohortsize,
       tau=NaN, accrual=NaN, tite.dist=NaN, accrual.dist=NaN, 
       add.args=list(alp.prior=phi, bet.prior=1-phi))
}
\arguments{
\item{nsimu}{the total number of trials to be simulated. The default value is 5000.}

\item{design}{option for selecting different designs, including CFO, aCFO, TITE-CFO, TITE-aCFO, fCFO,
f-aCFO, bCFO, and b-aCFO. Specifically, bCFO refers to the benchmark CFO, and b-aCFO
denotes the benchmark aCFO.}

\item{phi}{the target DLT rate.}

\item{p.true}{the true DLT rates under the different dose levels.}

\item{ncohort}{the total number of cohorts.}

\item{init.level}{the dose level assigned to the first cohort. The default value \code{init.level} is 1.}

\item{cohortsize}{the sample size in each cohort.}

\item{tau}{maximal assessment window size. 'NaN' should be assigned if the design without late-oneset outcomes.}

\item{accrual}{the accrual rate, i.e., the number of patients accrued in tau time. 'NaN' should be assigned
if the design without late-oneset outcomes.}

\item{tite.dist}{the distribution of the time to DLT events. \code{tite.dist=1} corresponds to a uniform distribution,
\code{tite.dist=2} corresponds to a Weibull distribution, and \code{tite.dist=3} corresponds to a
log-logistic distribution. 'NaN' should be assigned if the design without late-oneset outcomes.}

\item{accrual.dist}{the distribution of the arrival times of patients. When \code{tite.dist=1}, it corresponds to all
patients in each cohort arriving simultaneously at a given accrual rate. When \code{tite.dist=2},
it corresponds to a uniform distribution, and when \code{tite.dist=3}, it corresponds to an
exponential distribution. 'NaN' should be assigned if the design without late-oneset outcomes.}

\item{add.args}{additional parameters, usually set as list(alp.prior=phi, bet.prior=1-phi) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}
}
\value{
the \code{ CFO.oc() } function returns basic setup of ($simu.setup and $p.true) and the operating
characteristics ($selPercent,$nTox,$nPatients and $simu.oc) of the design. \cr
The operating characteristics includes \cr
(1) selection percentage at each dose level ($selPercent) \cr
(2) number of toxicity observed at each dose level ($nTox) \cr
(3) number of patients treated at each dose level ($nPatients) \cr
(4) percentage of correct selection of the MTD ($simu.oc$MTDSel) \cr
(5) percentage of patients allocated to the MTD ($simu.oc$MTDAllo) \cr
(6) percentage of selecting a dose above the MTD ($simu.oc$overSel) \cr
(7) percentage of allocating patients at dose levels above the MTD ($simu.oc$overAllo) \cr
(8) percentage of the patients suffering DLT ($simu.oc$averDLT) \cr
(9) Average trial duration if trials with late-onset toxicities ($simu.oc$averDur)
}
\description{
Obtain the operating characteristics of the CFO-type and aCFO-type designs for multiple simulations.
}
\details{
The operating characteristics of CFO-type and aCFO-type designs are generated by simulating trials under the
pre-specified true toxicity probabilities of the investigational doses. The choice of which design to execute
is determined by setting the \code{design} parameter. Some parameters (\code{tau}, \code{accrual},
\code{tite.dist}, and \code{accrual.dist}) need to be set only when running a design that can handle late-onset
toxicities; otherwise, they default to NaN.\cr
Additionally, in the example, we set \code{nsimu=100} for testing time considerations. In reality, \code{nsimu}
is typically set to 5000 to ensure the accuracy of the results.
}
\examples{
## setting
nsimu <- 100; phi <- 0.2; ncohort <- 12; cohortsize <- 3; init.level <- 1
p.true <-c(0.01, 0.05, 0.10, 0.14, 0.20, 0.26, 0.34)
add.args=list(alp.prior=phi, bet.prior=1-phi)
tau <- 3; accrual <- 6; tite.dist <- 2; accrual.dist <- 1

## get the operating characteristics for 100 simulations using the CFO design
CFO.oc (nsimu, design='CFO', phi, p.true, ncohort, init.level, cohortsize,
       tau=NaN, accrual=NaN, tite.dist=NaN, accrual.dist=NaN, add.args)
## get the operating characteristics for 100 simulations using the aCFO design
CFO.oc (nsimu, design='aCFO', phi, p.true, ncohort, init.level, cohortsize,
       tau=NaN, accrual=NaN, tite.dist=NaN, accrual.dist=NaN, add.args)
## get the operating characteristics for 100 simulations using the TITE-CFO design
CFO.oc (nsimu, design='TITE-CFO', phi, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, add.args)
## get the operating characteristics for 100 simulations using the TITE-aCFO design
CFO.oc (nsimu, design='TITE-aCFO', phi, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, add.args)
## get the operating characteristics for 100 simulations using the fCFO design
CFO.oc (nsimu, design='fCFO', phi, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, add.args)
## get the operating characteristics for 100 simulations using the f-aCFO design
CFO.oc (nsimu, design='f-aCFO', phi, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, add.args)
## get the operating characteristics for 100 simulations using the bCFO design
CFO.oc (nsimu, design='bCFO', phi, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, add.args)
## get the operating characteristics for 100 simulations using the b-aCFO design
CFO.oc (nsimu, design='b-aCFO', phi, p.true, ncohort, init.level, cohortsize,
        tau, accrual, tite.dist, accrual.dist, add.args)
}
