% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO.simu.R
\name{CFO.simu}
\alias{CFO.simu}
\title{Find the maximum tolerated dose (MTD) for a single Calibration-Free Odds (CFO) or accumulative CFO (aCFO) trial.}
\usage{
CFO.simu(phi, p.true, ncohort, init.level=1, cohortsize=3, 
       add.args=list(alp.prior=phi, bet.prior=1-phi), accumulation = FALSE, seed=NULL)
}
\arguments{
\item{phi}{the target DLT rate.}

\item{p.true}{the true DLT rates under the different dose levels.}

\item{ncohort}{the total number of cohorts.}

\item{init.level}{the dose level assigned to the first cohort. The default value \code{init.level} is 1.}

\item{cohortsize}{the sample size in each cohort. The default value \code{cohortsize} is 3.}

\item{add.args}{additional parameters, usually set as list(alp.prior=phi, bet.prior=1-phi) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{accumulation}{set \code{accumulation=FALSE} to conduct the CFO design; set \code{accumulation=TRUE} to conduct the
aCFO design.}

\item{seed}{an integer to set as the seed of the random number generator for reproducible results, the default is set to NULL.}
}
\value{
The \code{CFO.simu} function returns a list object comprising the following components:
\itemize{
\item{target: }{the target DLT rate.}
\item{p.true: }{the actual DLT rates under different dose levels.}
\item{MTD: }{the selected MTD.}
\item{dose.list: }{the list that includes the dose level assigned to each cohort.}
\item{DLT.ns: }{the total number of DLTs for all dose levels.}
\item{dose.ns: }{the total number of patients for all dose levels}
\item{tover.doses: }{the over-toxicity status for all dose levels (1 for yes).}
\item{DLT.list: }{The DLT outcome observed at each subject.}
}
}
\description{
Use this function to find the maximum tolerated dose (MTD) for a single Calibration-Free Odds (CFO) or accumulative CFO (aCFO) trial.
}
\details{
The \code{CFO.simu()} function is designed to determine the Maximum Tolerated Dose (MTD) for a single CFO or aCFO
trial. If \code{accumulation = FALSE}, this trial corresponds to the CFO design. If \code{accumulation = TRUE}, it
corresponds to the aCFO design. \cr
Given the toxicity outcomes from previous cohorts, each cohort is sequentially assigned to the most suitable dose
level based on the CFO or aCFO decision rule. Early stopping criteria are incorporated into the CFO or aCFO design
to ensure patient safety and benefit. If there is substantial evidence indicating that the current dose level
exhibits excessive toxicity (\eqn{\Pr(p_C > \phi|x_C, m_C \geq 3) > 0.95}), we exclude the current dose level as
well as higher dose levels from the trial. Upon the predefined maximum sample size is reached or the lowest dose
level is overly toxicity, the experiment is concluded, and the MTD is determined using isotonic regression.
}
\examples{
phi <- 0.2; ncohort <- 12; cohortsize <- 3
p.true <-c(0.01, 0.05, 0.10, 0.14, 0.20, 0.26, 0.34)
## find the MTD for a single CFO trial
CFO.simu(phi, p.true, ncohort, init.level=1, cohortsize=3,
         add.args=list(alp.prior=phi, bet.prior=1-phi), accumulation = FALSE)
## find the MTD for a single aCFO trial
CFO.simu(phi, p.true, ncohort, init.level=1, cohortsize=3,
         add.args=list(alp.prior=phi, bet.prior=1-phi), accumulation = TRUE)
}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066.
}
\author{
Jialu Fang
}
