% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFO2d.next.R
\name{CFO2d.next}
\alias{CFO2d.next}
\title{CFO2d.next function}
\usage{
CFO2d.next(phi, tys, tns, cur, alp.prior = 0.1, bet.prior = 0.9, seed = NULL)
}
\arguments{
\item{phi}{The efficacy threshold.}

\item{tys}{A matrix of the total number of successes for each dose combination.}

\item{tns}{A matrix of the total number of patients for each dose combination.}

\item{cur}{A vector of the current dose indices in the horizontal and vertical direction.}

\item{alp.prior}{Parameter alpha for the prior beta distribution.}

\item{bet.prior}{Parameter beta for the prior beta distribution.}

\item{seed}{Optional; an integer to set as the seed of the random number generator for reproducible results.}
}
\value{
A vector with the next dose indices in the horizontal and vertical directions.
}
\description{
This function is used to determine the next dose level in a two-dimensional
dose-finding clinical trial setting. It is based on the Continual Feedback On (CFO) method.
}
\examples{
tns <- matrix(c(3, 3, 3, 0,
                0, 0, 6, 0,
                0, 0, 0, 0), 
              nrow = 3, ncol = 4, byrow = TRUE)
tys <- matrix(c(0, 0, 1, 0,
                0, 0, 2, 0,
                0, 0, 0, 0), 
              nrow = 3, ncol = 4, byrow = TRUE)
cur <- c(2,3)
CFO2d.next(0.3, tys, tns, cur, seed = 1)

}
