% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateonset.next.R
\name{lateonset.next}
\alias{lateonset.next}
\title{Determination of the dose level for next cohort in the CFO-type and aCFO-type designs with late-onset toxicities}
\usage{
lateonset.next(curDose, phi, tau, impute.method, enter.times, dlt.times,
       current.t, accumulation, doses, tover.doses=c(), simu=FALSE,
       add.args=list(alp.prior=phi, bet.prior=1-phi))
}
\arguments{
\item{curDose}{the current dose level.}

\item{phi}{the target DLT rate.}

\item{tau}{maximal assessment window size}

\item{impute.method}{the imputing method for handling pending DLT data. \code{impute.method = 'frac'} corresponds to
fractional framework. \code{impute.method = 'TITE'} corresponds to time-to-event framework.
\code{impute.method = 'No'} implies no use of any imputing method, corresponding to the
benchmark CFO and benchmark aCFO designs.}

\item{enter.times}{the time at which each subject existing in the trial enters the trial.}

\item{dlt.times}{the time to DLT for each subject existing in the trial.  If no DLT occurs for a certain subject,
\code{dlt.times} is set to 0.}

\item{current.t}{the current time.}

\item{accumulation}{set \code{accumulation=FALSE} to conduct the CFO-type design; set \code{accumulation=TRUE} to
conduct the aCFO-type design.}

\item{doses}{the dose level for each subject existing in the trial.}

\item{tover.doses}{whether the dose level (from the first to last dose level) is over-toxic or not.
The value is set as 1 if the dose level is overly toxicity; otherwise, it is set to 0.
It should be predetermined. If not predetermined, it is set to NaN and is assigned values of 0
for each dose level.}

\item{simu}{whether simulation or not, if \code{simu=TRUE}, \code{lateonset.next()} also return \code{tover.doses}.}

\item{add.args}{additional parameters, usually set as list(alp.prior=phi, bet.prior=1-phi) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}
}
\value{
The \code{lateonset.next()} function returns the recommended dose level for treating the next cohort of
patients ($curDose). if \code{simu=TRUE}, \code{lateonset.next()} also return a vector indicating whether the
dose level (from the first to last dose level) is over-toxic or not ($tover.doses).
}
\description{
Propose the next dose level in the CFO-type and aCFO-type designs with late-onset toxicities, specifically, including
Time-to-event CFO (TITE-CFO) design, fractional CFO (fCFO) design, benchmark CFO design, TITE-aCFO design, f-aCFO
design and benchmark aCFO design.
}
\details{
Late-onset outcomes commonly occur in phase I trials involving targeted agents or immunotherapies. As a
result, the TITE framework and fractional framework serve as two imputation methods to handle pending data
related to late-onset outcomes. This approach extends the original designs to integrate time information
for delayed outcomes, leading to the development of TITE-CFO, fCFO, TITE-aCFO, and f-aCFO designs. \cr
In the TITE framework context, an assumption about the distribution of time to DLT must be pre-specified,
whereas the fractional framework does not require justification for a specific distribution of the time to
DLT. Consequently, fCFO and f-aCFO adapt to a more diverse range of scenarios.\cr
The function \code{lateonset.next()} also provides the option to set \code{impute.method = "No"} to execute
the benchmark CFO and aCFO design. These two methods await complete observation of toxicity outcomes for
the previous cohorts before determining the next dose assignment. This enhances precision but comes at the
expense of a prolonged trial duration.
}
\note{
The \code{tover.doses()} should be predetermined. It can be calculated based on the number of patients enrolled
($tns) and the number of DLTs observed ($tys), following the overdose control rule. If not predetermined, it is
set to NaN and is assigned default values of 0 for each dose level.
}
\examples{
## Given the parameters for the function, the unit for time-related parameters is in months. 
## The parameter generation follows the following guidelines: 
## 1) the assessment window ([0, $tau]) covers three months
## 2) the accrual rate is set at two patients per month. 
## 3) the arrival times of patients are distributed uniformly.
## 4) the time to DLT events is simulated using a Weibull distribution, with 50\% of these events 
##    occurring in the first half of the assessment window.
phi<-0.2
add.args=list(alp.prior=phi, bet.prior=1-phi)
enter.times<-c(0,0.082,0.343,0.554,1.18,1.88,2.15,2.68,2.74,3.30,3.99,
               4.52,5.51,5.93,6.18,6.68,7.05,7.92)
dlt.times<-c(0,0,0,0,0,0,0,0,0,0.934,0,0,0,0,0,1.5,0.6962,0)
current.t<-8.134413
doses<-c(1,1,1,2,2,2,3,3,3,4,4,4,3,3,3,4,4,4)
tover.doses<-c(0,0,0,0,0,0,0)
## determine the dose level for the next cohort using the TITE-CFO design
lateonset.next(curDose=4, phi, tau=3, impute.method="TITE", enter.times, dlt.times, current.t, 
               accumulation = FALSE, doses, tover.doses, simu=FALSE, add.args)
## determine the dose level for the next cohort using the TITE-aCFO design
lateonset.next(curDose=4, phi, tau=3, impute.method="TITE", enter.times, dlt.times, current.t, 
               accumulation = TRUE, doses, tover.doses, simu=FALSE, add.args)
## determine the dose level for the next cohort using the f-CFO design
lateonset.next(curDose=4, phi, tau=3, impute.method="frac", enter.times, dlt.times, current.t, 
               accumulation = FALSE, doses, tover.doses, simu=FALSE, add.args)
## determine the dose level for the next cohort using the f-aCFO design
lateonset.next(curDose=4, phi, tau=3, impute.method="frac", enter.times, dlt.times, current.t, 
               accumulation = TRUE, doses, tover.doses, simu=FALSE, add.args)
## determine the dose level for the next cohort using the benchmark CFO design
lateonset.next(curDose=4, phi, tau=3, impute.method="No", enter.times, dlt.times, current.t, 
               accumulation = FALSE, doses, tover.doses, simu=FALSE, add.args)
## determine the dose level for the next cohort using the benchmark aCFO design
lateonset.next(curDose=4, phi, tau=3, impute.method="No", enter.times, dlt.times, current.t, 
               accumulation = TRUE, doses, tover.doses, simu=FALSE, add.args)

}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066. \cr
Jin, H., & Yin, G. (2023). Time‐to‐event calibration‐free odds design: A robust efficient design for
phase I trials with late‐onset outcomes. \emph{Pharmaceutical Statistics}. \cr
Yin, G., Zheng, S., & Xu, J. (2013). Fractional dose-finding methods with late-onset toxicity in
phase I clinical trials. \emph{Journal of Biopharmaceutical Statistics}, 23(4), 856-870.
}
\author{
Jialu Fang
}
