% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateonset.simu.R
\name{lateonset.simu}
\alias{lateonset.simu}
\title{Find the maximum tolerated dose (MTD) for a single trial using the CFO-type and aCFO-type designs with late-onset toxicities.}
\usage{
lateonset.simu(phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist, accrual.dist, 
       design, init.dose=1, add.args=list(alp.prior=phi, bet.prior=1-phi), seed=NULL)
}
\arguments{
\item{phi}{the target DLT rate.}

\item{p.true}{the true DLT rates under the different dose levels.}

\item{tau}{maximal assessment window size.}

\item{cohortsize}{the sample size in each cohort.}

\item{ncohort}{the total number of cohorts.}

\item{accrual}{the accrual rate, i.e., the number of patients accrued in tau time.}

\item{tite.dist}{the distribution of the time to DLT events. \code{tite.dist=1} corresponds to a uniform distribution,
\code{tite.dist=2} corresponds to a Weibull distribution, and \code{tite.dist=3} corresponds to a
log-logistic distribution.}

\item{accrual.dist}{the distribution of the arrival times of patients. When \code{tite.dist=1}, it corresponds to all
patients in each cohort arriving simultaneously at a given accrual rate. When \code{tite.dist=2},
it corresponds to a uniform distribution, and when \code{tite.dist=3}, it corresponds to an
exponential distribution.}

\item{design}{option for selecting different designs, including TITE-CFO, TITE-aCFO, fCFO, f-aCFO, bCFO, and
b-aCFO. Specifically, bCFO refers to the benchmark CFO, and b-aCFO denotes the benchmark aCFO.}

\item{init.dose}{the dose level assigned to the first cohort. The default value \code{init.level} is 1.}

\item{add.args}{additional parameters, usually set as list(alp.prior=phi, bet.prior=1-phi) by default. \code{alp.prior}
and \code{bet.prior} represent the parameters of the prior distribution for the true DLT rate at
any dose level. This prior distribution is specified as Beta( \code{alpha.prior}, \code{beta.prior}).}

\item{seed}{an integer to set as the seed of the random number generator for reproducible results, the default is set to NULL.}
}
\value{
The \code{lateonset.simu()} function returns a list object comprising the following components: the target DLT
rate ($target), the actual DLT rates under different dose levels ($p.true), the selected MTD ($MTD),
the list that includes the dose level assigned to each cohort($dose.list), the total number of DLTs and
patients for all dose levels ($DLT.ns and $dose.ns), and the duration of the trial in months ($total.time).
}
\description{
Use this function to find the maximum tolerated dose (MTD) for the CFO-type and aCFO-type designs with late-onset toxicities,
specifically, including Time-to-event CFO (TITE-CFO) design, fractional CFO (fCFO), benchmark CFO, TITE-aCFO design,
the f-aCFO design and benchmark aCFO design.
}
\details{
The \code{lateonset.simu()} function is developed for determining the MTD in a single trial using CFO-type and
aCFO-type designs with late-onset toxicities. Specifically, it includes the TITE-CFO design, fCFO design, benchmark
CFO design, TITE-aCFO design, and f-aCFO design. The \code{design} parameter allows for selecting different designs.
Please note that the results returned by \code{lateonset.simu()} represent a single trial outcome. In practical
applications, multiple simulations are required to ascertain the MTD, and this can be achieved using the
\code{CFO.oc()} function.
}
\examples{
phi <- 0.2; ncohort <- 12; cohortsize <- 3
p.true <-c(0.01, 0.05, 0.10, 0.14, 0.20, 0.26, 0.34)
 tau <- 3; accrual <- 6; tite.dist <- 2; accrual.dist <- 1
## find the MTD for a single TITE-CFO trial
lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist, accrual.dist, 
                design='TITE-CFO', init.dose=1, add.args=list(alp.prior=phi, bet.prior=1-phi))
## find the MTD for a single TITE-aCFO trial
lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist, accrual.dist, 
                design='TITE-aCFO', init.dose=1, add.args=list(alp.prior=phi, bet.prior=1-phi))
## find the MTD for a single fCFO trial
lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist, accrual.dist, 
                design='fCFO', init.dose=1, add.args=list(alp.prior=phi, bet.prior=1-phi))
## find the MTD for a single f-aCFO trial
lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist, accrual.dist, 
                design='f-aCFO', init.dose=1, add.args=list(alp.prior=phi, bet.prior=1-phi))
## find the MTD for a single benchmark CFO trial
lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist, accrual.dist, 
                design='bCFO', init.dose=1, add.args=list(alp.prior=phi, bet.prior=1-phi))
## find the MTD for a single benchmark aCFO trial
lateonset.simu (phi, p.true, tau, cohortsize, ncohort, accrual, tite.dist, accrual.dist, 
                design='b-aCFO', init.dose=1, add.args=list(alp.prior=phi, bet.prior=1-phi))
}
\references{
Jin, H., & Yin, G. (2022). CFO: Calibration-free odds design for phase I/II clinical trials.
\emph{Statistical Methods in Medical Research}, 31(6), 1051-1066. \cr
Jin, H., & Yin, G. (2023). Time‐to‐event calibration‐free odds design: A robust efficient design for
phase I trials with late‐onset outcomes. \emph{Pharmaceutical Statistics}. \cr
Yin, G., Zheng, S., & Xu, J. (2013). Fractional dose-finding methods with late-onset toxicity in
phase I clinical trials. \emph{Journal of Biopharmaceutical Statistics}, 23(4), 856-870.
}
\author{
Jialu Fang
}
